// src/formBuilder/types.ts
export type DataType = "string" | "number" | "boolean" | "integer" | "array" | "object" | "null"

interface MatchType {
  types: Array<DataType>
  widget?: string
  field?: string
  format?: string
  $ref?: boolean
  enum?: boolean
}

export type CardComponentType = FunctionComponent<{
  parameters: CardComponentPropsType
  onChange: (newParams: CardComponentPropsType) => void
  mods?: Mods
}>

// an abstract input type
interface FormInputType {
  displayName: string
  // given a data and ui schema, determine if the object is of this input type
  matchIf: Array<MatchType>
  // allowed keys for ui:options
  possibleOptions?: Array<string>
  defaultDataSchema: {
    [key: string]: any
  }
  defaultUiSchema: {
    [key: string]: any
  }
  // the data schema type
  type: DataType
  // inputs on the preview card
  cardBody: CardComponentType
  // inputs for the modal
  modalBody?: CardComponentType
}

export interface DataOptions {
  title: string
  type?: string
  description?: string
  $ref?: string
  default?: string | number
}

export interface ModLabels {
  formNameLabel?: string
  formDescriptionLabel?: string
  objectNameLabel?: string
  displayNameLabel?: string
  descriptionLabel?: string
  inputTypeLabel?: string
  addElementLabel?: string
  addSectionLabel?: string
}

// optional properties that can add custom features to the form builder
export interface Mods {
  customFormInputs?: {
    [key: string]: FormInputType
  }
  components?: {
    add?: (properties?: { [key: string]: any }) => ReactElement | ReactElement[] | []
  }
  tooltipDescriptions?: {
    add?: string
    cardObjectName?: string
    cardDisplayName?: string
    cardDescription?: string
    cardInputType?: string
    cardSectionObjectName?: string
    cardSectionDisplayName?: string
    cardSectionDescription?: string
  }
  labels?: ModLabels
  showFormHead?: boolean
  deactivatedFormInputs?: Array<string>
  newElementDefaultDataOptions?: DataOptions
  newElementDefaultUiSchema?: { [key: string]: any }
}

export type FormInput = FormInputType

export interface InitParameters {
  categoryHash?: { [key: string]: string }
}
// ...
export interface AddFormObjectParametersType {
  schema: { [key: string]: any }
  uischema: { [key: string]: any }
  mods?: Mods
  onChange: (schema: { [key: string]: any }, uischema: { [key: string]: any }) => any
  definitionData: { [key: string]: any }
  definitionUi: { [key: string]: any }
  index?: number
  categoryHash: { [key: string]: string }
}